requirements-dev.txt
ruff
pytype
prospector
mypy
django-stubs
djangorestframework-stubs
pre-commit
black

